Code: Decouple Compilers from Margin. * Code Refacroting. * Analysis Fixes.
