+++ EditorExtensions/Commands/Css/CssSaveListener.cs
@@ -78,7 +78,7 @@ namespace MadsKristensen.EditorExtensions
             byte[] gzipContent = Compress(content);
             File.WriteAllBytes(gzipFile, gzipContent);
             ProjectHelpers.AddFileToActiveProject(gzipFile);
-            MarginBase.AddFileToProject(file, gzipFile);
+            FileHelpers.AddFileToProject(file, gzipFile);
         }
 
         public static byte[] Compress(string text)
