+++ EditorExtensions/Compilers/LESS/LessCompiler.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
+using System.Web.Helpers;
 using MadsKristensen.EditorExtensions.Helpers;
 using Microsoft.CSS.Core;
 
@@ -13,6 +16,7 @@ namespace MadsKristensen.EditorExtensions
         private static readonly Regex _endingCurlyBraces = new Regex(@"}\W*}|}", RegexOptions.Compiled);
         private static readonly Regex _linesStartingWithTwoSpaces = new Regex("(\n( *))", RegexOptions.Compiled);
         private static readonly Regex _errorParsingPattern = new Regex(@"^(?<message>.+) in (?<fileName>.+) on line (?<line>\d+), column (?<column>\d+):$", RegexOptions.Multiline);
+        private static readonly Regex _sourceMapInCss = new Regex(@"\/\*#([^*]|[\r\n]|(\*+([^*/]|[\r\n])))*\*\/", RegexOptions.Multiline);
 
         protected override string ServiceName
         {
@@ -62,6 +66,8 @@ namespace MadsKristensen.EditorExtensions
                         targetFile: targetFileName,
                         oldBasePath: sourceFileName
                     );
+
+                    resultSource = UpdateSourceMapUrls(resultSource, targetFileName);
                 }
                 catch (Exception ex)
                 {
@@ -72,6 +78,48 @@ namespace MadsKristensen.EditorExtensions
             Logger.Log(message);
 
             return resultSource;
+        }
+
+        private static string UpdateSourceMapUrls(string content, string compiledFileName)
+        {
+            if (!WESettings.GetBoolean(WESettings.Keys.LessSourceMaps) || !File.Exists(compiledFileName))
+                return content;
+
+            string sourceMapFilename = compiledFileName + ".map";
+
+            if (!File.Exists(sourceMapFilename))
+                return content;
+
+            var updatedFileContent = GetUpdatedSourceMapFileContent(compiledFileName, sourceMapFilename);
+
+            if (updatedFileContent == null)
+                return content;
+
+            FileHelpers.WriteFile(updatedFileContent, sourceMapFilename);
+            FileHelpers.AddFileToProject(compiledFileName, sourceMapFilename);
+
+            return UpdateSourceLinkInCssComment(content, FileHelpers.RelativePath(compiledFileName, sourceMapFilename));
+        }
+
+        private static string GetUpdatedSourceMapFileContent(string cssFileName, string sourceMapFilename)
+        {
+            // Read JSON map file and deserialize.
+            dynamic jsonSourceMap = Json.Decode(File.ReadAllText(sourceMapFilename));
+
+            if (jsonSourceMap == null)
+                return null;
+
+            jsonSourceMap.sources = ((IEnumerable<dynamic>)jsonSourceMap.sources).Select(s => FileHelpers.RelativePath(cssFileName, s));
+            jsonSourceMap.names = new List<dynamic>(jsonSourceMap.names);
+            jsonSourceMap.file = Path.GetFileName(cssFileName);
+
+            return Json.Encode(jsonSourceMap);
+        }
+
+        private static string UpdateSourceLinkInCssComment(string content, string sourceMapRelativePath)
+        {   // Fix sourceMappingURL comment in CSS file with network accessible path.
+            return _sourceMapInCss.Replace(content,
+                string.Format(CultureInfo.CurrentCulture, "/*# sourceMappingURL={0} */", sourceMapRelativePath));
         }
     }
 }
